+++ compiler/test/dotty/tools/scripting/BashScriptsTests.scala
@@ -13,27 +13,27 @@ import vulpix.TestConfiguration
 
 import ScriptTestEnv.*
 
-/** Verifies correct handling of command line arguments by `dist/bin/scala` and `dist/bin/scalac`.
- *   +. arguments following a script path must be treated as script arguments
- *   +. preserve script command line arguments.
- *   +. prevent SCALA_OPTS in build environment from infecting tests, via 'SCALA_OPTS= ' prefix
- *   +. test scripts must not throw execptions or exit with nonzero.
- */
+/** Verifies correct handling of command line arguments by `dist/bin/scala` and
+  * `dist/bin/scalac`. +. arguments following a script path must be treated as
+  * script arguments +. preserve script command line arguments. +. prevent
+  * SCALA_OPTS in build environment from infecting tests, via 'SCALA_OPTS= '
+  * prefix +. test scripts must not throw execptions or exit with nonzero.
+  */
 object BashScriptsTests:
   lazy val argsfile = createArgsFile() // avoid problems caused by drive letter
   def testFiles = scripts("/scripting")
 
-  @AfterClass def cleanup: Unit = {
+  @AfterClass def cleanup: Unit =
     val af = argsfile.toFile
-    if (af.exists) {
-      af.delete()
-    }
-  }
+    if af.exists then af.delete()
   printf("osname[%s]\n", osname)
   printf("uname[%s]\n", ostypeFull)
   printf("using JAVA_HOME=%s\n", envJavaHome)
   printf("using SCALA_HOME=%s\n", envScalaHome)
-  printf("first 5 PATH entries:\n%s\n", adjustedPathEntries.take(5).mkString("\n"))
+  printf(
+    "first 5 PATH entries:\n%s\n",
+    adjustedPathEntries.take(5).mkString("\n")
+  )
   printf("scala path:  [%s]\n", scalaPath)
   printf("scalac path: [%s]\n", scalacPath)
 
@@ -44,15 +44,21 @@ object BashScriptsTests:
     "arg  3:[-repl]",
     "arg  4:[-run]",
     "arg  5:[-script]",
-    "arg  6:[-debug]",
+    "arg  6:[-debug]"
   )
   val testScriptArgs = Seq(
-    "a", "b", "c", "-repl", "-run", "-script", "-debug"
+    "a",
+    "b",
+    "c",
+    "-repl",
+    "-run",
+    "-script",
+    "-debug"
   )
   val showArgsScript = testFiles.find(_.getName == "showArgs.sc").get.absPath
 
   def testFile(name: String): String =
-    val file = testFiles.find(_.getName == name) match {
+    val file = testFiles.find(_.getName == name) match
       case Some(f) =>
         val ff = f.absPath
         printf("test file [%s] is [%s]\n", name, ff)
@@ -60,21 +66,25 @@ object BashScriptsTests:
       case None =>
         printf("test file [%s] not found!\n", name)
         name.absPath
-    }
     file
 
-  val Seq(envtestSc, envtestScala) = Seq("envtest.sc", "envtest.scala").map { testFile(_) }
+  val Seq(envtestSc, envtestScala) = Seq("envtest.sc", "envtest.scala").map {
+    testFile(_)
+  }
 
   // create command line with given options, execute specified script, return stdout
   def callScript(tag: String, script: String, keyPre: String): String =
     val keyArg = s"$keyPre=$tag"
     printf("pass tag [%s] via [%s] to script [%s]\n", tag, keyArg, script)
-    val cmd: String = Seq("SCALA_OPTS= ", scalaPath, keyArg, script).mkString(" ")
+    val cmd: String =
+      Seq("SCALA_OPTS= ", scalaPath, keyArg, script).mkString(" ")
     printf("cmd: [%s]\n", cmd)
     val (validTest, exitCode, stdout, stderr) = bashCommand(cmd)
-    stderr.filter { !_.contains("Inappropriate ioctl") }.foreach { System.err.printf("stderr [%s]\n", _) }
+    stderr.filter { !_.contains("Inappropriate ioctl") }.foreach {
+      System.err.printf("stderr [%s]\n", _)
+    }
     stdout.mkString("\n")
-
+end BashScriptsTests
 
 class BashScriptsTests:
   import BashScriptsTests.*
@@ -86,7 +96,7 @@ class BashScriptsTests:
   @Test def verifyScJProperty =
     val tag = "World1"
     val stdout = callScript(tag, envtestSc, s"-J-Dkey")
-    assertEquals( s"Hello $tag", stdout)
+    assertEquals(s"Hello $tag", stdout)
 
   /* verify that `dist/bin/scala` correctly passes args to the jvm via -J-D for script envtest.scala */
   @Test def verifyScalaJProperty =
@@ -108,50 +118,58 @@ class BashScriptsTests:
 
   /* verify that `dist/bin/scala` can set system properties via -D when executing compiled script via -jar envtest.jar */
   @Test def saveAndRunWithDProperty =
-    val commandline = Seq("SCALA_OPTS= ", scalaPath.relpath, "-save", envtestScala.relpath).mkString(" ")
+    val commandline =
+      Seq("SCALA_OPTS= ", scalaPath.relpath, "-save", envtestScala.relpath)
+        .mkString(" ")
     val (_, _, _, _) = bashCommand(commandline) // compile jar, discard output
-    val testJar = testFile("envtest.jar") // jar is created by the previous bashCommand()
-    if (testJar.isFile){
+    val testJar = testFile(
+      "envtest.jar"
+    ) // jar is created by the previous bashCommand()
+    if testJar.isFile then
       printf("compiled envtest.scala to %s\n", testJar.norm)
-    } else {
-      sys.error(s"error: unable to compile envtest.scala to ${testJar.norm}")
-    }
+    else sys.error(s"error: unable to compile envtest.scala to ${testJar.norm}")
 
     val tag = "World5"
-    val commandline2 = Seq("SCALA_OPTS= ", scalaPath.relpath, s"-Dkey=$tag", testJar.relpath)
+    val commandline2 =
+      Seq("SCALA_OPTS= ", scalaPath.relpath, s"-Dkey=$tag", testJar.relpath)
     printf("cmd[%s]\n", commandline2.mkString(" "))
-    val (validTest, exitCode, stdout, stderr) = bashCommand(commandline2.mkString(" "))
+    val (validTest, exitCode, stdout, stderr) = bashCommand(
+      commandline2.mkString(" ")
+    )
     assertEquals(s"Hello $tag", stdout.mkString("/n"))
 
   /* verify `dist/bin/scalac` non-interference with command line args following script name */
   @Test def verifyScalacArgs =
-    val commandline = (Seq("SCALA_OPTS= ", scalacPath, "-script", showArgsScript) ++ testScriptArgs).mkString(" ")
+    val commandline = (Seq(
+      "SCALA_OPTS= ",
+      scalacPath,
+      "-script",
+      showArgsScript
+    ) ++ testScriptArgs).mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     if verifyValid(validTest) then
       var fail = false
       printf("\n")
       for (line, expect) <- stdout zip expectedOutput do
         printf("expected: %-17s\nactual  : %s\n", expect, line)
-        if line != expect then
-          fail = true
+        if line != expect then fail = true
 
-      if fail then
-        assert(stdout == expectedOutput)
+      if fail then assert(stdout == expectedOutput)
 
   /* verify `dist/bin/scala` non-interference with command line args following script name */
   @Test def verifyScalaArgs =
-    val commandline = (Seq("SCALA_OPTS= ", scalaPath, showArgsScript) ++ testScriptArgs).mkString(" ")
+    val commandline =
+      (Seq("SCALA_OPTS= ", scalaPath, showArgsScript) ++ testScriptArgs)
+        .mkString(" ")
     val (validTest, exitCode, stdout, stderr) = bashCommand(commandline)
     if verifyValid(validTest) then
       var fail = false
       printf("\n")
       for (line, expect) <- stdout zip expectedOutput do
         printf("expected: %-17s\nactual  : %s\n", expect, line)
-        if line != expect then
-          fail = true
+        if line != expect then fail = true
 
-      if fail then
-        assert(stdout == expectedOutput)
+      if fail then assert(stdout == expectedOutput)
 
   /*
    * verify that scriptPath.sc sees a valid script.path property,
@@ -161,37 +179,56 @@ class BashScriptsTests:
   @Test def verifyScriptPathProperty =
     val scriptFile = testFiles.find(_.getName == "scriptPath.sc").get
     val expected = s"${scriptFile.getName}"
-    printf("===> verify valid system property script.path is reported by script [%s]\n", scriptFile.getName)
+    printf(
+      "===> verify valid system property script.path is reported by script [%s]\n",
+      scriptFile.getName
+    )
     printf("calling scriptFile: %s\n", scriptFile)
     val (validTest, exitCode, stdout, stderr) = bashCommand(scriptFile.absPath)
     if verifyValid(validTest) then
       stdout.foreach { printf("stdout: [%s]\n", _) }
       stderr.foreach { printf("stderr: [%s]\n", _) }
       val valid = stdout.exists { _.endsWith(expected) }
-      if valid then printf("# valid script.path reported by [%s]\n", scriptFile.getName)
-      assert(valid, s"script ${scriptFile.absPath} did not report valid script.path value")
+      if valid then
+        printf("# valid script.path reported by [%s]\n", scriptFile.getName)
+      assert(
+        valid,
+        s"script ${scriptFile.absPath} did not report valid script.path value"
+      )
 
   /*
    * verify SCALA_OPTS can specify an @argsfile when launching a scala script in `dist/bin/scala`.
    */
   @Test def verifyScalaOpts =
     val scriptFile = testFiles.find(_.getName == "classpathReport.sc").get
-    printf("===> verify SCALA_OPTS='@argsfile' is properly handled by `dist/bin/scala`\n")
+    printf(
+      "===> verify SCALA_OPTS='@argsfile' is properly handled by `dist/bin/scala`\n"
+    )
     val envPairs = List(("SCALA_OPTS", s"@$argsfile"))
-    val (validTest, exitCode, stdout, stderr) = bashCommand(scriptFile.absPath, envPairs)
-    printf("stdout: %s\n", stdout.mkString("\n","\n",""))
+    val (validTest, exitCode, stdout, stderr) =
+      bashCommand(scriptFile.absPath, envPairs)
+    printf("stdout: %s\n", stdout.mkString("\n", "\n", ""))
     if verifyValid(validTest) then
       val expected = s"${workingDirectory.norm}"
       // stdout might be polluted with an ANSI color prefix, so be careful
-      val cwdline = stdout.find( _.trim.matches(".*cwd: .*") ).getOrElse("")
+      val cwdline = stdout.find(_.trim.matches(".*cwd: .*")).getOrElse("")
       printf("cwdline  [%s]\n", cwdline)
       printf("expected[%s]\n", expected)
       val valid = cwdline.endsWith(expected)
-      if (!valid) then
+      if !valid then
         stdout.foreach { printf("stdout[%s]\n", _) }
         stderr.foreach { printf("stderr[%s]\n", _) }
-      if valid then printf(s"\n===> success: classpath begins with %s, as reported by [%s]\n", workingDirectory, scriptFile.getName)
-      assert(valid, s"script ${scriptFile.absPath} did not report valid java.class.path first entry")
+      if valid then
+        printf(
+          s"\n===> success: classpath begins with %s, as reported by [%s]\n",
+          workingDirectory,
+          scriptFile.getName
+        )
+      assert(
+        valid,
+        s"script ${scriptFile.absPath} did not report valid java.class.path first entry"
+      )
+  end verifyScalaOpts
 
   /*
    * verify that individual scripts can override -save with -nosave (needed to address #13760).
@@ -199,35 +236,53 @@ class BashScriptsTests:
   @Test def sqlDateTest =
     val scriptBase = "sqlDateError"
     val scriptFile = testFiles.find(_.getName == s"$scriptBase.sc").get
-    val testJar = testFile(s"$scriptBase.jar") // jar should not be created when scriptFile runs
+    val testJar = testFile(
+      s"$scriptBase.jar"
+    ) // jar should not be created when scriptFile runs
     val tj = Paths.get(testJar).toFile
     if tj.isFile then tj.delete() // discard residual debris from previous test
-    printf("===> verify '-save' is cancelled by '-nosave' in script hashbang.`\n")
-    val (validTest, exitCode, stdout, stderr) = bashCommand(s"SCALA_OPTS=-save ${scriptFile.absPath}")
-    printf("stdout: %s\n", stdout.mkString("\n","\n",""))
+    printf(
+      "===> verify '-save' is cancelled by '-nosave' in script hashbang.`\n"
+    )
+    val (validTest, exitCode, stdout, stderr) = bashCommand(
+      s"SCALA_OPTS=-save ${scriptFile.absPath}"
+    )
+    printf("stdout: %s\n", stdout.mkString("\n", "\n", ""))
     if verifyValid(validTest) then
       // the script should print '1969-12-31' or '1970-01-01', depending on time zone
       // stdout can be polluted with an ANSI color prefix, in some test environments
       val valid = stdout.mkString("").matches(""".*\d{4}-\d{2}-\d{2}.*""")
-      if (!valid) then
+      if !valid then
         stdout.foreach { printf("stdout[%s]\n", _) }
         stderr.foreach { printf("stderr[%s]\n", _) }
-      if valid then printf(s"\n===> success: scripts can override -save via -nosave\n")
-      assert(valid, s"script ${scriptFile.absPath} reported unexpected value for java.sql.Date ${stdout.mkString("\n")}")
-      assert(!testJar.exists,s"unexpected, jar file [$testJar] was created")
-
+      if valid then
+        printf(s"\n===> success: scripts can override -save via -nosave\n")
+      assert(
+        valid,
+        s"script ${scriptFile.absPath} reported unexpected value for java.sql.Date ${stdout.mkString("\n")}"
+      )
+      assert(!testJar.exists, s"unexpected, jar file [$testJar] was created")
+  end sqlDateTest
 
   /*
    * verify -e println("yo!") works.
    */
   @Test def verifyCommandLineExpression =
-    printf("===> verify -e <expression> is properly handled by `dist/bin/scala`\n")
+    printf(
+      "===> verify -e <expression> is properly handled by `dist/bin/scala`\n"
+    )
     val expected = "9"
     val expression = s"println(3*3)"
     val cmd = s"bin/scala -e $expression"
-    val (validTest, exitCode, stdout, stderr) = bashCommand(s"""bin/scala -e '$expression'""")
+    val (validTest, exitCode, stdout, stderr) = bashCommand(
+      s"""bin/scala -e '$expression'"""
+    )
     val result = stdout.filter(_.nonEmpty).mkString("")
     printf("stdout: %s\n", result)
-    printf("stderr: %s\n", stderr.mkString("\n","\n",""))
+    printf("stderr: %s\n", stderr.mkString("\n", "\n", ""))
     if verifyValid(validTest) then
-      assert(result.contains(expected), s"expression [$expression] did not send [$expected] to stdout")
+      assert(
+        result.contains(expected),
+        s"expression [$expression] did not send [$expected] to stdout"
+      )
+end BashScriptsTests
